+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/params/index.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, it } from 'vitest'
-import { paramsObject, queryParameter, requestParameter, requestParamsArray } from '.'
+import { paramsObject, queryParameter, requestParameter } from '.'
 
 // Test run
 // pnpm vitest run ./src/generator/zod-openapi-hono/openapi/route/params/index.test.ts
@@ -11,9 +11,6 @@ describe('request module barrel file exports', () => {
 
   it('should export requestParameter', () => {
     expect(typeof requestParameter).toBe('function')
-  })
-  it('should export requestParamsArray', () => {
-    expect(typeof requestParamsArray).toBe('function')
   })
   it('should export queryParameter', () => {
     expect(typeof queryParameter).toBe('function')
