+++ packages/hono-takibi/src/generator/zod-openapi-hono/openapi/route/params/index.ts
@@ -1,4 +1,3 @@
 export { paramsObject } from './params-object.js'
 export { queryParameter } from './query-parameter.js'
 export { requestParameter } from './request-parameter.js'
-export { requestParamsArray } from './request-params-array.js'
